+++ src/plugins/intel_gna/src/memory/gna_memory_state.cpp
@@ -107,7 +107,7 @@ InferenceEngine::Blob::CPtr GNAVariableState::GetState() const {
         auto buffer = result_blob->buffer().as<float*>();
         auto new_gna_ptr = static_cast<int16_t*>(state->gna_ptr);
 
-        for (int i = 0; i < elements; i++) {
+        for (size_t i = 0; i < elements; i++) {
             buffer[i] = new_gna_ptr[i] / scale_factor;
         }
